Adding two things.
First, an algorithm to take all the extra single hits leftover
from a clustering algorithm and make them into their own clusters.

Next, an algorithm to merge single hits into clusters that they
are very close to.

Other changes are minor things to the makefiles, a script to run the merging,
and the linkdef file.